Der gemmes i database, nå der genereres en rapport

+++ src/main/java/dk/kea/projekt3_gruppe6_bilabonnement/Service/DashboardService.java
@@ -18,7 +18,7 @@ public class DashboardService {
     }
 
     //Antal udlejde biler
-    public long seAntalUdlejdeBiler() {
+    public int seAntalUdlejdeBiler() {
         return bilRepository.findByStatus("Udlejet").size();
     }
